requirements.txt
zhipuai
gradio
jinja2
pyyaml